Move ProfileSetting.vue to oj/views/user
- Remove oj/views/settings/*
- Fix import path of ProfileSetting.vue from Header.vue and Problem.vue
- Fix ProfileSetting access path from routes.js
