+++ frontend/src/pages/oj/router/routes.js
@@ -10,12 +10,12 @@ import {
   Problem,
   ProblemList,
   ResetPassword,
+  ProfileSetting,
   SubmissionDetails,
   SubmissionList
 } from '../views'
 
 import * as Contest from '@oj/views/contest'
-import * as Setting from '@oj/views/setting'
 
 export default [
   {
@@ -115,22 +115,10 @@ export default [
     ]
   },
   {
+    name: 'profile-setting',
     path: '/setting',
-    component: Setting.Settings,
-    children: [
-      {
-        name: 'default-setting',
-        path: '',
-        meta: { requiresAuth: true, title: 'Default Settings' },
-        component: Setting.ProfileSetting
-      },
-      {
-        name: 'profile-setting',
-        path: 'profile',
-        meta: { requiresAuth: true, title: 'Profile Settings' },
-        component: Setting.ProfileSetting
-      }
-    ]
+    meta: { title: 'Profile Settings' },
+    component: ProfileSetting
   },
   {
     path: '*',
