--- frontend/src/pages/oj/views/setting/index.js
@@ -1,4 +1,0 @@
-const Settings = () => import(/* webpackChunkName: "setting" */ './Settings.vue')
-const ProfileSetting = () => import(/* webpackChunkName: "setting" */ './children/ProfileSetting.vue')
-
-export { Settings, ProfileSetting }
