+++ frontend/src/pages/oj/views/user/ProfileSetting.vue
@@ -1,69 +1,71 @@
 <template>
-  <div class="setting-main font-bold">
-    <div class="section-title logo-title font-bold">
-      <h2>Setting</h2>
-    </div>
-    <div style="display:flex; flex-direction:row; margin-bottom:12px;" class="profileSetting">
-      <b-form>
-        <h4 class="columnName font-bold">Preferred Language</h4>
-          <b-row class="mb-4">
-          <b-form-select class="setting-select" v-model="formProfile.language" :options="languages" :selected="formProfile.language"></b-form-select>
-          </b-row>
-          <b-row class="mb-4">
-            <b-button class="setting-btn" variant="success" @click="updateLanguage">
-              <b-spinner v-if="loading.btnLanguage" small></b-spinner> Change Language
-            </b-button>
-          </b-row>
-      </b-form>
-      <b-form>
-        <h4 class="columnName font-bold">Major</h4>
-          <b-row class="mb-4">
-          <b-form-select class="setting-select" v-model="formProfile.major" :options="majors" :selected="formProfile.major"></b-form-select>
-          </b-row>
-          <b-row>
-            <b-button class="setting-btn" variant="success" @click="updateMajor">
-              <b-spinner v-if="loading.btnMajor" small></b-spinner> Change Major
-            </b-button>
-          </b-row>
-      </b-form>
-    </div>
-    <div style="display:flex; flex-direction:row;" class="accountSetting">
-      <b-form>
-        <h4 class="columnName font-bold">Change Password</h4>
-        <b-form-group class="setting-label" label= "Current Password" prop="old_password">
-          <b-form-input class="setting-input" type="password" v-model="formPassword.old_password" required></b-form-input>
-        </b-form-group>
-        <b-form-group class="setting-label" label="New Password" prop="new_password">
-          <b-form-input class="setting-input" type="password" v-model="formPassword.new_password" required></b-form-input>
-        </b-form-group>
-        <b-form-group class="setting-label" label="Confirm New Password" prop="again_password">
-          <b-form-input class="setting-input" type="password" v-model="formPassword.again_password" required></b-form-input>
-        </b-form-group>
-        <b-form-group v-if="visible.tfaRequired" label="Two Factor Auth" prop="tfa_coded">
-          <b-form-input class="setting-input" type="password" v-model="formPassword.again_password" required></b-form-input>
-        </b-form-group>
-        <b-button class="setting-btn" variant="success" @click="changePassword">
-          <b-spinner v-if="loading.btnPassword" small></b-spinner> Change Password
-        </b-button>
-      </b-form>
-      <b-form>
-        <h4 class="columnName font-bold">Change Email</h4>
-        <b-form-group class="setting-label" label="Old Email" >
-          <b-form-input class="setting-input" v-model="formEmail.old_email" disabled></b-form-input>
-        </b-form-group>
-        <b-form-group class="setting-label" label="New Email" prop="new_email">
-          <b-form-input class="setting-input" type="email" v-model="formEmail.new_email"></b-form-input>
-        </b-form-group>
-        <b-form-group class="setting-label" label="Current Password" prop="password">
-          <b-form-input class="setting-input" type="password" v-model="formEmail.password"></b-form-input>
-        </b-form-group>
-        <b-form-group v-if="visible.tfaRequired" label="Two Factor Auth" prop="tfa_coded">
-          <b-form-input type="text" v-model="formEmail.tfa_code"></b-form-input>
-        </b-form-group>
-        <b-button class="setting-btn" variant="success" @click="changeEmail">
-          <b-spinner v-if="loading.btnEmail" small></b-spinner> Change Email
-        </b-button>
-      </b-form>
+  <div class="container">
+    <div class="setting-main font-bold">
+      <div class="section-title logo-title font-bold">
+        <h2>Setting</h2>
+      </div>
+      <div style="display:flex; flex-direction:row; margin-bottom:12px;" class="profileSetting">
+        <b-form>
+          <h4 class="columnName font-bold">Preferred Language</h4>
+            <b-row class="mb-4">
+            <b-form-select class="setting-select" v-model="formProfile.language" :options="languages" :selected="formProfile.language"></b-form-select>
+            </b-row>
+            <b-row class="mb-4">
+              <b-button class="setting-btn" variant="success" @click="updateLanguage">
+                <b-spinner v-if="loading.btnLanguage" small></b-spinner> Change Language
+              </b-button>
+            </b-row>
+        </b-form>
+        <b-form>
+          <h4 class="columnName font-bold">Major</h4>
+            <b-row class="mb-4">
+            <b-form-select class="setting-select" v-model="formProfile.major" :options="majors" :selected="formProfile.major"></b-form-select>
+            </b-row>
+            <b-row>
+              <b-button class="setting-btn" variant="success" @click="updateMajor">
+                <b-spinner v-if="loading.btnMajor" small></b-spinner> Change Major
+              </b-button>
+            </b-row>
+        </b-form>
+      </div>
+      <div style="display:flex; flex-direction:row;" class="accountSetting">
+        <b-form>
+          <h4 class="columnName font-bold">Change Password</h4>
+          <b-form-group class="setting-label" label= "Current Password" prop="old_password">
+            <b-form-input class="setting-input" type="password" v-model="formPassword.old_password" required></b-form-input>
+          </b-form-group>
+          <b-form-group class="setting-label" label="New Password" prop="new_password">
+            <b-form-input class="setting-input" type="password" v-model="formPassword.new_password" required></b-form-input>
+          </b-form-group>
+          <b-form-group class="setting-label" label="Confirm New Password" prop="again_password">
+            <b-form-input class="setting-input" type="password" v-model="formPassword.again_password" required></b-form-input>
+          </b-form-group>
+          <b-form-group v-if="visible.tfaRequired" label="Two Factor Auth" prop="tfa_coded">
+            <b-form-input class="setting-input" type="password" v-model="formPassword.again_password" required></b-form-input>
+          </b-form-group>
+          <b-button class="setting-btn" variant="success" @click="changePassword">
+            <b-spinner v-if="loading.btnPassword" small></b-spinner> Change Password
+          </b-button>
+        </b-form>
+        <b-form>
+          <h4 class="columnName font-bold">Change Email</h4>
+          <b-form-group class="setting-label" label="Old Email" >
+            <b-form-input class="setting-input" v-model="formEmail.old_email" disabled></b-form-input>
+          </b-form-group>
+          <b-form-group class="setting-label" label="New Email" prop="new_email">
+            <b-form-input class="setting-input" type="email" v-model="formEmail.new_email"></b-form-input>
+          </b-form-group>
+          <b-form-group class="setting-label" label="Current Password" prop="password">
+            <b-form-input class="setting-input" type="password" v-model="formEmail.password"></b-form-input>
+          </b-form-group>
+          <b-form-group v-if="visible.tfaRequired" label="Two Factor Auth" prop="tfa_coded">
+            <b-form-input type="text" v-model="formEmail.tfa_code"></b-form-input>
+          </b-form-group>
+          <b-button class="setting-btn" variant="success" @click="changeEmail">
+            <b-spinner v-if="loading.btnEmail" small></b-spinner> Change Email
+          </b-button>
+        </b-form>
+      </div>
     </div>
   </div>
 </template>
@@ -74,6 +76,7 @@ import utils from '@/utils/utils'
 import { types } from '@/store'
 
 export default {
+  name: 'ProfileSetting',
   data () {
     return {
       loading: {
@@ -237,7 +240,23 @@ export default {
 <style lang="less" scoped>
   @font-face {
     font-family: Manrope_bold;
-    src: url('../../../../../fonts/Manrope-Bold.ttf');
+    src: url('../../../../fonts/Manrope-Bold.ttf');
+  }
+  .container {
+    min-width: 800px;
+    margin: auto;
+    padding: 0;
+  }
+  .setting-main {
+    position: relative;
+    margin: 10px 40px;
+    padding-bottom: 20px;
+    .setting-content {
+      margin-left: 20px;
+    }
+    .mini-container {
+      width: 500px;
+    }
   }
   .profileSetting {
     display:flex;
